hopefully doing feature scaling right?

+++ utils.py
@@ -3,7 +3,7 @@ import tensorflow as tf
 
 def merge(images, size):
     h, w = images.shape[1], images.shape[2]
-    img = np.zeros((h * size[0], w * size[1], 3))
+    img = np.zeros((h * size[0], w * size[1], 3), dtype=np.uint8)
 
     for idx, image in enumerate(images):
         i = idx % size[1]
